requirements.txt
appdirs==1.4.4
astroid==2.4.2
distlib==0.3.1
filelock==3.0.12
future==0.18.2
importlib-metadata==2.0.0
importlib-resources==3.0.0
isort==5.6.4
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
numpy==1.19.2
pandas==1.1.4
pbr==5.5.0
Pillow==8.0.0
pylint==2.6.0
six==1.15.0
scikit-image==0.17.2
stevedore==3.2.2
tensorboard==2.4.0
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
torchsummary==1.5.1
typed-ast==1.4.1
virtualenv==20.0.34
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wrapt==1.12.1
zipp==3.3.0
